requirements.txt
Django==4.2.28
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.7.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
gunicorn==25.1.0
whitenoise==6.12.0
dj-database-url==3.1.2
